+++ lib/core/share/share_service.dart
@@ -4,5 +4,7 @@ import '../../features/voucher/domain/entities/voucher_receipt.dart';
 abstract interface class ShareService {
   Future<void> shareVoucherReceipt(VoucherReceipt receipt);
 
+  Future<void> shareVoucherReceipts(List<VoucherReceipt> receipts);
+
   Future<void> shareReportExport(ReportExport export);
 }